feat: totalSales employee added

+++ src/Providers/tables.js
@@ -1,11 +1,16 @@
-import { createContext, useState } from "react";
+import { createContext, useContext, useState } from "react";
 import { api } from "../Services/api";
+import { AdminContext } from "./admin";
+import { toast } from "react-toastify";
 
 export const TablesContext = createContext();
 
 export const TablesProvider = ({ children }) => {
   const [tables, setTables] = useState([]);
-  const userId = localStorage.getItem("@id");
+  const userId = localStorage.getItem("@userId");
+  const id = localStorage.getItem("@id");
+  const { updateEmployee } = useContext(AdminContext);
+  const token = localStorage.getItem("@token");
 
   function addTable(table) {
     const tableExists = tables.find(
@@ -22,16 +27,26 @@ export const TablesProvider = ({ children }) => {
     }
   }
 
-  function removeTable(tableId) {
-    api.delete(`/tables/${tableId}`);
-    syncTables();
+  async function removeTable(tableId) {
+    const isAdmin = await updateEmployee(id, tableId);
+
+    if (isAdmin) {
+      api.delete(`/tables/${tableId}`);
+      toast.success("Mesa finalizada.");
+    }
+    setTimeout(() => {
+      syncTables();
+    }, 300);
   }
 
   async function syncTables() {
+    const tempUserId = localStorage.getItem("@userId");
+    const tempId = localStorage.getItem("@id");
     await api
-      .get(`/tables?userId=${userId}`)
-      .then((res) => setTables(res.data));
-    ;
+      .get(`/tables?userId=${token ? tempId : tempUserId}`)
+      .then((res) => {
+        setTables(res.data);
+      });
   }
 
   function updateTable(numberTable, products) {
